requirement.txt
asgiref==3.8.1
astroid==3.3.8
dill==0.3.9
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
isort==5.13.2
mccabe==0.7.0
platformdirs==4.3.6
PyJWT==2.10.1
pylint==3.3.3
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
